backend/requirements.txt
fastapi
uvicorn
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic
pydantic-settings
python-dotenv==1.0.1
requests
beautifulsoup4
lxml
httpx
alembic
pdfplumber
python-dateutil
cloudscraper
pytesseract
pdf2image
Pillow
playwright==1.41.2
openbb
# openbb-sectors==1.1.2 removed due to availability issues


